Update to 1.2.1
**Prior to 1.2.1:**

All passwords generated with QtPass's built-in password generator are possibly predictable and enumerable by hackers. The generator used libc's random(), seeded with srand(msecs), where msecs is not the msecs since 1970 (not that that'd be secure anyway), but rather the msecs since the last second. This means there are only 1000 different sequences of generated passwords.

**1.2.1:**

All passwords that have been generated with QtPass prior to 1.2.1 should be regenerated and changed.